Seg post processing
The purpose of this script is to provide an example code, which mainly accomplishes two tasks. The first task is to conduct a post-processing of the segmentation project. If there are multiple segmentation instances, it will carefully separate them. The second task is to check whether the segmented instances intersect with the predefined boundaries, that is, to measure the distance between them and the boundaries.